removed old simulation script
